Web/Pequod Package Provider/Challenge/Private/src/requirements.txt
argon2==0.1.10
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-Login==0.6.3
greenlet==3.0.3
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.21
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.1
gunicorn==21.2.0
packaging==24.0